# Test performance of find InboxId 