Performance

```
 LOG  Found inboxId from address in 93.42199993133545ms
 LOG  Found inboxId from address in 87.63558292388916ms
 LOG  Found inboxId from address in 87.89387500286102ms
 LOG  Found inboxId from address in 93.37329196929932ms
 LOG  Found inboxId from address in 93.67287492752075ms
 LOG  Found inboxId from address in 90.73575007915497ms
 LOG  Found inboxId from address in 90.45954203605652ms
```